Tentative code layout using roles
The idea is that the application should create its own provider which
builds on an existing provider, pulling the TFA role and building the
around modifier.

Additional routes, etc. should be created by the role on BUILD method,
which appears to work.

Also the application should provide the code to store and access the token
(next step)